+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 3.0.0
 
 - Add chart of the scores (line chart adding a point to GIoU and MAHA curves on every change.)
-- Add checkbox "Show chart" in the main menu. 
+- Add checkbox "Show chart" in the main menu.
 
 ## 2.2.0